requirements.txt
Django==4.1.10
django-ckeditor==6.5.1
django-environ==0.9.0
django-extensions==3.2.1
django-js-asset==2.2.0
djangorestframework==3.14.0
ruff==0.3.1
black==24.2.0
environ==1.0
psycopg2==2.9.9